+++ fe/package.json
@@ -17,8 +17,6 @@
     "cors": "^2.8.5",
     "dayjs": "^1.11.11",
     "framer-motion": "^11.1.9",
-    "dayjs": "^1.11.11",
-    "framer-motion": "^11.1.9",
     "google-maps": "^4.3.3",
     "google-maps-react": "^2.0.6",
     "localforage": "^1.10.0",
@@ -27,13 +25,11 @@
     "react-datepicker": "^6.9.0",
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
-    "react-dropzone": "^14.2.3",
     "react-icons": "^5.0.1",
+    "react-loader-spinner": "^6.1.6",
     "react-router-dom": "^6.18.0",
     "react-scripts": "5.0.1",
     "sort-by": "^1.2.0",
-    "styled-components": "^6.1.8",
-    "tailwindcss": "^3.4.3",
     "styled-components": "^6.1.8",
     "tailwindcss": "^3.4.3",
     "web-vitals": "^2.1.4"
@@ -61,9 +57,6 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@tailwindcss/forms": "^0.5.7"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.7"
