+++ fe/src/pages/EditProfileUI/EditProfileUI.jsx
@@ -8,74 +8,6 @@ import EditPhone from "../../components/EditProfileComponents/EditPhone";
 import UserBookingHistory from "../../components/UserBookingHistory/UserBookingHistory";
 import axios from "axios";
 
-// const EditProfile3 = () => {
-//   const [profile, setProfile] = useState(null);
-//   const [loading, setLoading] = useState(true);
-//   const [error, setError] = useState(null);
-//   const [user, setUser] = useState(null);
-  
-//   useEffect(() => {
-//     // const token = document.cookie.split(';').find(c => c.trim().startsWith('auth_token='));
-//     const token = localStorage.getItem("auth_token");
-
-//     // console.log("Token:", token);
-//     if (token) {
-//       const jwtToken = token.split("=")[1];
-//       axios
-//         .post("http://127.0.0.1:8000/api/decodetoken", { token: token })
-//         .then((response) => {
-//           setUser(response.data["data"]);
-          
-//         })
-//         .catch((error) => {
-//           alert("Error decoding JWT token:", error);
-//           setUser(null);
-//         });
-//         console.log ("user token",token);
-//     } else {
-//       setUser(null);
-//     }
-//   }, []);
-  
-//   useEffect(() => {
-//     const fetchProfile = async () => {
-//       try {
-//         const response = await axios.get(`http://127.0.0.1:8000/api/getusers/${user.userid}`);
-//         console.log("Response Data:", response.data); // Log the entire response object
-//         console.log("Response Data Email:", response.data.userid); // Log the entire response object
-//         setProfile(response.data);
-//         setLoading(false);
-//       } catch (error) {
-//         setError("Error fetching profile data.");
-//         console.error(error);
-//         setLoading(false);
-//       }
-//     };
-
-//     fetchProfile();
-//   }, []);
-//   useEffect(() => {
-//     const fetchProfile = async () => {
-//       try {
-//         const response = await axios.get("http://localhost/api/loadProfile.php", {
-//           params: {
-//             userid: 1 // Replace with the logged in user's id
-//           }
-//         });
-//         console.log("Response Data:", response.data); // Log the entire response object
-//         console.log("Response Data Email:", response.data.userid); // Log the entire response object
-//         setProfile(response.data);
-//         setLoading(false);
-//       } catch (error) {
-//         setError("Error fetching profile data.");
-//         console.error(error);
-//         setLoading(false);
-//       }
-//     };
-
-//     fetchProfile();
-//   }, []);
-
 const EditProfileUI = () => {
   const [profile, setProfile] = useState(null);
   const [loading, setLoading] = useState(true);
@@ -83,7 +15,44 @@ const EditProfileUI = () => {
   const [currentView, setCurrentView] = useState('profile'); // State to manage the current view
   const [editingName, setEditingName] = useState(false);
   const [editingPhone, setEditingPhone] = useState(false);
-  const { user } = useUser();
+  const [user, setUser] = useState(null);
+
+  useEffect(() => {
+    const token = localStorage.getItem("auth_token");
+    if (token) {
+      axios
+        .post("http://127.0.0.1:8000/api/decodetoken", { token: token })
+        .then((response) => {
+          setUser(response.data["data"]);
+          console.log("RESPONSE DATA: ", response.data["data"]);
+        })
+        .catch((error) => {
+          alert("Error decoding JWT token:", error);
+          setUser(null);
+        });
+    } else {
+      setUser(null);
+    }
+  }, []);
+
+  useEffect(() => {
+    const fetchProfile = async () => {
+      if (!user) return; // Exit if user is not set
+      try {
+        const response = await axios.get(`http://127.0.0.1:8000/api/getusers/${user.userid}`);
+        console.log("Response Data:", response.data); // Log the entire response object
+        console.log("Response Data UserID:", response.data.userid); // Log the entire response object
+        setProfile(response.data);
+        setLoading(false);
+      } catch (error) {
+        setError("Error fetching profile data.");
+        console.error(error);
+        setLoading(false);
+      }
+    };
+
+    fetchProfile();
+  }, [user, editingName, editingPhone]); // Add user as a dependency here
 
   // Dummy data
   const dummyProfileData = {
@@ -96,54 +65,13 @@ const EditProfileUI = () => {
     ]
   };
 
-  // useEffect(() => {
-  //   // Simulate API call
-  //   setTimeout(() => {
-  //     setProfile(dummyProfileData);
-  //     setLoading(false);
-  //   }, 1000); // Simulating 1 second delay
-  // }, []);
-
   useEffect(() => {
-    const fetchProfile = async () => {
-      try {
-        const response = await axios.get(`http://127.0.0.1:8000/api/getusers/${user.userid}`);
-        console.log("Response Data:", response.data); // Log the entire response object
-        console.log("Response Data Email:", response.data.userid); // Log the entire response object
-        setProfile(response.data);
-        setLoading(false);
-      } catch (error) {
-        setError("Error fetching profile data.");
-        console.error(error);
-        setLoading(false);
-      }
-    };
-
-    fetchProfile();
+    // Simulate API call
+    setTimeout(() => {
+      setProfile(dummyProfileData);
+      setLoading(false);
+    }, 1000); // Simulating 1 second delay
   }, []);
-  useEffect(() => {
-    const fetchProfile = async () => {
-      try {
-        const response = await axios.get("http://localhost/api/loadProfile.php", {
-          params: {
-            userid: 1 // Replace with the logged in user's id
-          }
-        });
-        console.log("Response Data:", response.data); // Log the entire response object
-        console.log("Response Data Email:", response.data.userid); // Log the entire response object
-        setProfile(response.data);
-        setLoading(false);
-      } catch (error) {
-        setError("Error fetching profile data.");
-        console.error(error);
-        setLoading(false);
-      }
-    };
-
-    fetchProfile();
-  }, []);
-
-
 
   if (loading) return <p>Loading...</p>;
   if (error) return <p>{error}</p>;
@@ -158,9 +86,9 @@ const EditProfileUI = () => {
 
   const renderProfileDetails = () => (
     <div>
-      <div className="edit-section" >
+      <div className="edit-section">
         <h2 className="edit-section-header" style={{ fontFamily: 'Poppins' }}>USER DETAILS</h2>
-        <div className="edit-container" >
+        <div className="edit-container">
           <div
             className="edit-detail-container"
             style={{ backgroundColor: 'white', boxShadow: 'rgba(0, 0, 0, 0.2) 0px 1px 3px 0px, rgba(0, 0, 0, 0.14) 0px 1px 1px -2px, rgba(0, 0, 0, 0.12) 0px 2px 1px 0px', padding: '20px', borderRadius: '8px', marginBottom: '20px', cursor: 'pointer', position: 'relative' }}
@@ -204,30 +132,13 @@ const EditProfileUI = () => {
             )}
           </div>
         </div>
-     
-        {/* <h2 className="edit-section-header" style={{ fontFamily: 'Poppins' }}>PAYMENT METHODS</h2>
-        <div className="edit-container">
-          {profile.paymentMethods.map((method, index) => (
-            <div
-              key={index}
-              className="edit-detail-container"
-              style={{ backgroundColor: 'white', boxShadow: 'rgba(0, 0, 0, 0.2) 0px 1px 3px 0px, rgba(0, 0, 0, 0.14) 0px 1px 1px -2px, rgba(0, 0, 0, 0.12) 0px 2px 1px 0px', padding: '20px', borderRadius: '8px', marginBottom: '20px', cursor: 'pointer', position: 'relative' }}
-              onMouseEnter={(e) => { e.currentTarget.style.backgroundColor = 'lightblue' }}
-              onMouseLeave={(e) => { e.currentTarget.style.backgroundColor = 'white' }}
-            >
-              <div className="edit-detail-label" style={{ textAlign: 'left' }}>{method.type}</div>
-              <div className="edit-detail" style={{ textAlign: 'left' }}>{method.details}</div>
-              <span className="edit-edit-text">Edit</span>
-            </div>
-          ))}
-        </div> */}
       </div>
     </div>
   );
 
   return (
-    <div className="edit-profile-container" >
-      <div className="edit-sidebar" >
+    <div className="edit-profile-container">
+      <div className="edit-sidebar">
         <div
           className={`edit-sidebar-item ${currentView === 'profile' ? 'active' : ''}`}
           onClick={() => setCurrentView('profile')}
@@ -239,7 +150,6 @@ const EditProfileUI = () => {
         <div
           className={`edit-sidebar-item ${currentView === 'history' ? 'active' : ''}`}
           onClick={() => setCurrentView('history')}
-         
         >
           <MdHistory className="edit-sidebar-icon" />
           <span className="edit-sidebar-text">History</span>
@@ -247,7 +157,6 @@ const EditProfileUI = () => {
         <div
           className={`edit-sidebar-item ${currentView === 'reviews' ? 'active' : ''}`}
           onClick={() => setCurrentView('reviews')}
-         
         >
           <MdStar className="edit-sidebar-icon" />
           <span className="edit-sidebar-text">Reviews</span>
