+++ fe/src/pages/EditProfileUI/UserProfile.css
@@ -55,7 +55,7 @@
       radial-gradient(circle at bottom right, aquamarine 50%, transparent 50%);
     background-color: lightblue;
     box-shadow: rgba(0, 0, 0, 0.1) 0px 2px 4px;
-    height: 100%;
+    height: 60rem;
   }
