+++ fe/src/pages/PropertyManagementUI/components/cancelled.js
@@ -17,7 +17,7 @@ import {
   Button
 } from "@mui/material";
 import { Edit as EditIcon } from "@mui/icons-material";
-import EditReservationModal from "../modals/guests";
+import EditReservationModal from "../modals/EditReservationModal";
 
 const Cancelled = () => {
   const [open, setOpen] = useState(false);
