+++ fe/src/pages/PropertyManagementUI/components/checkout.js
@@ -17,7 +17,7 @@ import {
   Button
 } from "@mui/material";
 import { Edit as EditIcon } from "@mui/icons-material";
-import EditReservationModal from "../modals/guests";
+import EditReservationModal from "../modals/EditReservationModal";
 
 const Checkouts = ({data}) => {
   const [open, setOpen] = useState(false);
