+++ fe/src/pages/SearchAndFilter/components/MainContent.css
@@ -319,6 +319,7 @@ body {
     margin-bottom: 5px;
     margin-top: 2rem;
     margin-left: 0.3rem;
+    
   }
   
   .filter-button {
