+++ fe/src/pages/SearchAndFilter/components/Search.css
@@ -11,7 +11,6 @@
     box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
     border-radius: 8px;
     overflow: hidden;
-    margin: 10px;
     margin-right: 10px;
     padding: 10px;
     width: 64rem;
